pyproject.toml
[project]
name = "hackwashu-backend2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.120.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.38.0",
]
